requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.8.0
django-filter==2.4.0
django-querycount==0.7.0
django-rest-passwordreset==1.2.1
Pillow==8.4.0